requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-pptx==0.6.23
reportlab==4.0.7
pypdf==3.17.4
python-docx==1.1.0
gunicorn==21.2.0
